Create Zero Tolerance Plugin
This plugin was taken from Events.js and repurposed for the Marketplace Staff & Marketplace Rooms. It was added due to the fact that we currently have to use a rudimentary zero tolerance pastebin; I do plan to look into opening this command into all rooms, so that each room can have its list of zero tolerance users that they can keep track of easily and efficiently.